Module sunrise_kernel::devices::ioapic

82093AA I/O Advanced Programmable Interrupt Controller (IOAPIC) driver

The IO-APIC is used to dispatch external device and inter-process interruptions to the correct CPU.

The documentation for the IO-APIC can be found here.

Structs

IoApic
See module level documentation

IoApicInternal
Internal IO-APIC registers.

RedirectionEntry
Description of a Redirection Entry in the IO-APIC. Unlike IRQ pins of the 8259A, the notion of interrupt priority is completely unrelated to the position of the physical interrupt input signal on the APIC. Instead, software determines the vector (and therefore the priority) for each corresponding interrupt input signal. For each interrupt signal, the operating system can also specify the signal polarity (low active or high active), whether the interrupt is signaled as edges or levels, as well as the destination and delivery mode of the interrupt. The information in the redirection table is used to translate the corresponding interrupt pin information into an inter-APIC message.

Enums

DeliveryMode
Specifies how the APICs listed in the destination field should act upon reception of this signal. Note that certain Delivery Modes only operate as intended when used in conjunction with a specific trigger Mode.