Module sunrise_kernel::devices::lapic

Local APIC driver.

Specification can be found in Intel's System Programmer's Guide. The chapters used as reference here will be from this version of the guide.

Structs

Error
Contains the set of errors the LAPIC has encountered while running.

LocalApic
LocalApic driver.

LocalApicInternal
Local APIC Register structure.

LocalApicRegister
Local APIC Registers are 128-bit wide, with the 32 lower bits containing the actual register, and the top bits being reserved for future use.

LocalApicVersion
The version and associated metadata of a Local APIC are described by this struct.

LocalVector
Allows software to specify the manner in which the local interrupts are delivered to the processor core.

SpuriousInterrupt
See chapter 10.9: Spurious Interrupt

Enums

DeliveryMode
Specifies how the APICs listed in the destination field should act upon reception of this signal. Note that certain Delivery Modes only operate as intended when used in conjunction with a specific trigger Mode.

TimerMode
Selects the Timer Mode of the LVT Timer.