Struct sunrise_kernel::devices::pic::Pic

pub struct Pic {
    master: SpinLockIRQ<InternalPic>,
    slave: SpinLockIRQ<InternalPic>,
}

A master/slave PIC setup, as commonly found on IBM PCs.

Fields
master: SpinLockIRQ<InternalPic>

The master PIC.

slave: SpinLockIRQ<InternalPic>

The slave PIC, cascaded on line 2 of .master

Implementations

impl Pic

unsafe fn new() -> Pic

Creates a new PIC, and initializes it.

Interrupts will be mapped to IRQ [32..48]

Safety

This should only be called once! If called more than once, then both Pics instances will share the same underlying Pios, but different mutexes protecting them!

pub fn mask(&self, irq: u8)

Mask the given IRQ number. Will redirect the call to the right Pic device.

pub fn unmask(&self, irq: u8)

Unmask the given IRQ number. Will redirect the call to the right Pic device.

pub fn get_mask(&self) -> u16

Reads the PIC interrupt mask. Used for debug purposes.

LSB is irq 0, MSB is irq 15.

pub fn acknowledge(&self, irq: u8)

Acknowledges an IRQ, allowing the PIC to send a new IRQ on the next cycle.