Enum sunrise_kernel::error::KernelError

pub enum KernelError {
    NotImplemented {
        msg: &'static str,
        backtrace: Backtrace,
    },
    PhysicalMemoryExhaustion {
        backtrace: Backtrace,
    },
    VirtualMemoryExhaustion {
        backtrace: Backtrace,
    },
    InvalidAddress {
        address: usize,
        backtrace: Backtrace,
    },
    InvalidSize {
        size: usize,
        backtrace: Backtrace,
    },
    ProcessKilled {
        backtrace: Backtrace,
    },
    InvalidState {
        backtrace: Backtrace,
    },
    InvalidCombination {
        backtrace: Backtrace,
    },
    ExceedingMaximum {
        value: u64,
        maximum: u64,
        backtrace: Backtrace,
    },
    InvalidKernelCaps {
        kcap: u32,
        backtrace: Backtrace,
    },
    IpcError {
        backtrace: Backtrace,
    },
    WrongMappingFramesForTy {
        ty: MemoryType,
        backtrace: Backtrace,
    },
    InvalidMemState {
        address: VirtualAddress,
        ty: MemoryType,
        backtrace: Backtrace,
    },
    ReservedValue {
        backtrace: Backtrace,
    },
}

Kernel Error.

Used pretty much everywhere that an error can occur. Holds the reason of the error, and a backtrace of its origin, for debug.

When a KernelError must be propagated to userspace, i.e. a syscall failed, it must be converted to a UserspaceError.

Variants

NotImplemented

Fields of NotImplemented

msg: &'static strbacktrace: Backtrace

PhysicalMemoryExhaustion

Fields of PhysicalMemoryExhaustion

backtrace: Backtrace

VirtualMemoryExhaustion

Fields of VirtualMemoryExhaustion

backtrace: Backtrace

InvalidAddress

Fields of InvalidAddress

address: usizebacktrace: Backtrace

InvalidSize

Fields of InvalidSize

size: usizebacktrace: Backtrace

ProcessKilled

Fields of ProcessKilled

backtrace: Backtrace

InvalidState

Fields of InvalidState

backtrace: Backtrace

InvalidCombination

Fields of InvalidCombination

backtrace: Backtrace

ExceedingMaximum

Fields of ExceedingMaximum

value: u64maximum: u64backtrace: Backtrace

InvalidKernelCaps

Fields of InvalidKernelCaps

kcap: u32backtrace: Backtrace

IpcError

Fields of IpcError

backtrace: Backtrace

WrongMappingFramesForTy

Fields of WrongMappingFramesForTy

ty: MemoryTypebacktrace: Backtrace

InvalidMemState

Fields of InvalidMemState

address: VirtualAddressty: MemoryTypebacktrace: Backtrace

ReservedValue

Fields of ReservedValue

backtrace: Backtrace