Struct sunrise_kernel::error::UserspaceError

#[repr(transparent)]pub struct UserspaceError(u32);

Kernel syscall error codes.

When a syscall fails, it returns one of this values as a reason for the fail.

Implementations

impl KernelError

pub fn make_ret(self) -> u32

Transforms a KernelError into the encoding acceptable for a syscall return value.

pub fn from_syscall_ret(err: u32) -> KernelError

Turns a syscall return value into a Kernel Error.

pub fn from_description(err: u32) -> KernelError

Turns a kernel error description into a KernelError.

pub fn description(self) -> u32

Gets the underlying KernelError description.

impl KernelError

pub const InvalidKernelCaps: KernelError

Kernel capabilities are invalid.

pub const NotImplemented: KernelError

This function is not implemented.

pub const InvalidSize: KernelError

The size argument is invalid.

Generally means it's not properly aligned.

pub const InvalidAddress: KernelError

The passed address is invalid.

Generally means it is not page aligned.

pub const MemoryFull: KernelError

The virtual address space was exhausted.

pub const HandleTableFull: KernelError

The process' handle table is full.

pub const InvalidMemState: KernelError

The memory state is invalid for this action.

pub const InvalidMemPerms: KernelError

The memory permissions passed are wrong.

pub const InvalidMemRange: KernelError

Memory range is not at an expected location.

pub const InvalidThreadPriority: KernelError

Invalid thread priority. Thread priority should be within the range 0..=0x3F, and should be allowed in the kernel capabilities.

pub const InvalidProcessorId: KernelError

Invalid processor id. Processor ID should exist on the current machine and be allowed in the kernel capabilities.

pub const InvalidHandle: KernelError

Passed handle is invalid.

Either the handle passed is of the wrong type, or the handle number wasn't valid at all.

pub const CopyFromUserFailed: KernelError

Attempt to copy the userspace address failed.

pub const InvalidCombination: KernelError

The combination of argument is invalid.

pub const Timeout: KernelError

A timeout was reached.

pub const Canceled: KernelError

The syscall was cancelled through cancel_synchronization.

pub const ExceedingMaximum: KernelError

A size or address was given exceeding the maximum allowed value.

pub const InvalidEnum: KernelError

No enum variants match this integer value.

pub const NoSuchEntry: KernelError

The given entry does not exist.

pub const PortRemoteDead: KernelError

The remote part of the session was closed.

pub const InvalidState: KernelError

Attempted to do an operation that's invalid in the handle's current state.

pub const ReservedValue: KernelError

Attempted to use an unknown value, reserved for future use.