Module sunrise_kernel::frame_allocator

Physical memory manager.

This module can only allocate and free whole frames.

Re-exports

pub use self::physical_mem_region::PhysicalMemRegion;

pub use self::physical_mem_region::PhysicalMemRegionIter;

Modules

i386
Architecture specific-behaviour i386 implementation of the frame allocator.

physical_mem_region
PhysicalMemRegion

private
Private FrameAllocator API

Structs

FrameAllocator
The physical memory manager.

Traits

FrameAllocatorTrait
An arch-specific FrameAllocator must expose the following functions

Functions

init
Initialize the FrameAllocator by parsing the multiboot information and marking some memory areas as unusable

mark_frame_bootstrap_allocated
Marks a physical memory frame as already allocated Currently used during init when paging marks KernelLand frames as alloc'ed by bootstrap