Struct sunrise_kernel::i386::gdt::GdtManager

pub struct GdtManager {
    table_a: DescriptorTable,
    table_b: DescriptorTable,
    table_selector: bool,
}

Safety wrapper that manages the lifetime of GDT tables.

Although Intel's guide doesn't really say much about it, modifying a GDT "live" is probably a terrible idea. To work around this, the GdtManager keeps two copies of the DescriptorTable, one being the currently active one (loaded in the GDTR), and the other being where the changes to the GDT go to until they are committed.

When commit is called, the internal GDT and current GDTR are swapped.

This struct's implementation of Deref and DerefMut will always give a reference to the table currently not in use, so you can make modifications to it, and call commit afterwards.

Fields
table_a: DescriptorTable

One of the two tables.

table_b: DescriptorTable

One of the two tables.

table_selector: bool

The table currently pointed to by GDTR. 0 is table_a, 1 is table_b.

Implementations

impl GdtManager

pub fn commit(
&mut self,
new_cs: Option<SegmentSelector>,
new_ds: Option<SegmentSelector>,
new_es: Option<SegmentSelector>,
new_fs: Option<SegmentSelector>,
new_gs: Option<SegmentSelector>,
new_ss: Option<SegmentSelector>
)

Commit the changes in the currently unloaded table, and update segment registers.

Selectors

To make a segment register point to a new descriptor, pass Some(selector) to this function.

If None is passed, the register will be reloaded from its current value. This is what you want if you only updated the content of the descriptor. We always perform a reload of all registers to make sure they reflect the state of the GDT, in case the user modified it.