Module sunrise_kernel::i386::instructions::segmentation

Provides functions to read and write segment registers.

Functions

cs
Returns the current value of the code segment register.

ds
Read the value of the data segment register.

es
Read the value of the es segment register.

fs
Read the value of the fs segment register.

gs
Read the value of the gs segment register.

load_ds⚠
Reload data segment register.

load_es⚠
Reload es segment register.

load_fs⚠
Reload fs segment register.

load_gs⚠
Reload gs segment register.

load_ss⚠
Reload stack segment register.

set_cs⚠
Reload code segment register. Note this is special since we can not directly move to %cs. Instead we push the new segment selector and return value on the stack and use lretq to reload cs and continue at 1:.

ss
Read the value of the stack segment register.