Module sunrise_kernel::i386::multiboot

Multiboot Information

Gives access to the multiboot information structure created by our bootloader (GRUB).

Bootloader (GRUB) passed its address in $ebx when we started.

Our bootstrap had the job of copy-ing it to a PAGE_SIZE aligned address, map it, and passed it to us in $ebx.

When kernel initializes we store this address in BOOT_INFO, and we can then access it at any moment by calling get_boot_information.

Statics

BOOT_INFO
Stores the address of the multiboot.

Functions

get_boot_information
Get a pointer to the multiboot information structure.

init
Initializes the boot information module, allowing the get_boot_information functions to operate properly.

try_get_boot_information
Tries to get a pointer to the multiboot information structure.