Module sunrise_kernel::i386::process_switch

Arch-specific process switch functions

This modules describe low-level functions and structures needed to perform a process switch

Structs

ThreadHardwareContext
The hardware context of a paused thread. It contains just enough registers to get the thread running again.

Functions

first_schedule⚠
The function ret'd on, on a thread's first schedule - as setup by the prepare_for_first_schedule.

jump_to_entrypoint
Jumps to Userspace, and run a userspace program.

prepare_for_first_schedule⚠
Prepares the thread for its first schedule by writing default values at the start of the stack that will be loaded in the registers in schedule-in. See process_switch() documentation for more details.

process_switch⚠
Performs the process switch, switching from currently running process A, to process B.