Module sunrise_kernel::i386::structures::idt

Provides types for the Interrupt Descriptor Table and its entries.

Structs

EntryOptions
Represents the options field of an IDT entry.

ExceptionStackFrame
Represents the exception stack frame pushed by the CPU on exception entry.

Idt
An Interrupt Descriptor Table with 256 entries.

IdtEntry
An Interrupt Descriptor Table entry.

PageFaultErrorCode
Describes an page fault error code.

Enums

GateType
Represents the type of an IDT descriptor (called a gate).

Type Definitions

HandlerFunc
A handler function for an interrupt or an exception without error code.

HandlerFuncWithErrCode
A handler function for an exception that pushes an error code.

PageFaultHandlerFunc
A page fault handler function that pushes a page fault error code.