Struct sunrise_kernel::mem::PhysicalAddress

#[repr(transparent)]pub struct PhysicalAddress(pub usize);

Represents a Physical address

Implementations

impl PhysicalAddress

pub const fn addr(self) -> usize

Gets the address as a usize.

impl PhysicalAddress

pub fn checked_add(self, rhs: usize) -> Option<PhysicalAddress>

Tries to add an offset to a PhysicalAddress, returning None if this would cause an overflow.

This function does not return a KernelError, as it does not know whether the address or the size is the cause of the error.

pub fn check_aligned_to(self, alignment: usize) -> Result<(), KernelError>

Checks that this address meets the given alignment.

Errors

InvalidAddress: self is not aligned to alignment.

pub fn floor(self) -> PhysicalAddress

Rounds down to PAGE_SIZE.

pub fn ceil(self) -> PhysicalAddress

Rounds up PAGE_SIZE.