Module sunrise_kernel::paging::arch::i386::lands

i386 Virtual Memory Layout

This module describes the splitting of memory for the i386 architecture.

The layout for the 4GB address space is the following:

0x00000000 - 0xbfffffff:  3GB of virtual memory belonging to the user.
0xc0000000 - 0xffbfffff: ~1GB of virtual memory belonging to the kernel.
0xffc00000 - 0xffffffff:  4MB of virtual memory pointing to the page tables themselves.

Structs

KernelLand
The virtual memory belonging to kernel.

RecursiveTablesLand
The virtual memory pointing to active page tables by recursion.

UserLand
The virtual memory belonging to user.

Constants

DIRECTORY_RECURSIVE_ADDRESS
When paging is on, accessing this address loops back to the directory itself thanks to recursive mapping on directory's last entry.

KERNELLAND_END_TABLE
The index in page directory of the last table of KernelLand.

KERNELLAND_START_TABLE
The index in page directory of the first table of KernelLand.

USERLAND_END_TABLE
The index in page directory of the last table of UserLand.

USERLAND_START_TABLE
The index in page directory of the first table of UserLand.