Module sunrise_kernel::paging::hierarchical_table

Arch-independent traits for architectures that implement paging as a hierarchy of page tables

Structs

NoFlush
Flusher that doesn't flush.

SmartHierarchicalTable
This is just a wrapper for a pointer to a table. It enables us to do handle when it is dropped

Enums

PageState
A hierarchical paging is composed of entries. An entry can be in the following states:

Traits

HierarchicalEntry
A hierarchical paging is composed of entries. All entries implements the following trait

HierarchicalTable
A hierarchical paging is composed of tables. All tables must implement the following trait A table of entries, either the top-level directory or one of the page tables. A table is a parent table if its child are also tables, instead of regular pages.

InactiveHierarchyTrait
A trait implemented by inactive table hierarchies.

PagingCacheFlusher
Most implementations of paging have are accelerated with a cache that must be manually updated when changes to the page tables are made. The way we specify which part of the cache gets invalidated is arch-specific. We only provide the declaration for a flusher that our page tables can use.

TableHierarchy
A trait operating on a whole hierarchy of tables.