Function sunrise_kernel::panic::display_bsod

fn display_bsod()

Display the infamous "Blue Screen Of Death"

When the kernel panics, we blit an image to the screen to inform the user we have kernel panicked.

This function attempts to map the framebuffer, and copies BSOD_BMP to it.

It is designed to fail silently if mapping the framebuffer or parsing the BMP failed, as it should only be called from the panic handler, and the last thing we want at that time is more error handling.

Note that this function will write to the framebuffer with no regards to whether it was already mapped by another process. This is OK since we're panicking, and all processes should be halted by now.