Module sunrise_kernel::process

Process

Modules

capabilities
Process Capability handling

thread_local_storage
TLS manager

Structs

HandleTable
Holds the table associating userspace handle numbers to a kernel Handle.

ProcessCapabilities
Capabilities of a process.

ProcessStateData
Data related to the (user-visible) state the current process is in. The maternity is stored here to ensure there is no race condition between setting the state to Exited and adding threads to the maternity.

ProcessStruct
The struct representing a process. There's one for every process.

ThreadStateEvent
The underlying shared object of a [Weak].

ThreadStruct
The struct representing a thread. A process may own multiple threads.

Enums

Handle
A handle to a userspace-accessible resource.

ThreadState
The state of a thread.

Statics

NEXT_PROCESS_ID
Next available PID.

PROCESS_LIST
List of processes currently running on the system.