Function sunrise_kernel::scheduler::add_to_schedule_queue

pub fn add_to_schedule_queue(thread: Arc<ThreadStruct>)

Adds a thread at the end of the schedule queue, and changes its state to 'scheduled' Thread must be ready to be scheduled.

If the thread was already scheduled, this function is a Noop.

Panics

Panics if the thread's state was already "Scheduled"