Function sunrise_kernel::scheduler::internal_schedule

fn internal_schedule<'a, LOCK, GUARD>(
    lock: &'a LOCK, 
    remove_self: bool
) -> GUARD where
    LOCK: Lock<'a, GUARD>,
    GUARD: 'a, 

Internal impl of the process switch, used by schedule and unschedule.

See schedule function for documentation on how scheduling works.