Struct sunrise_kernel::sync::spin_lock_irq::SpinLockIRQ

pub struct SpinLockIRQ<T: ?Sized> {
    internal: SpinLock<T>,
}

SpinLock that disables IRQ.

Description

This type behaves like a spinlock from the Linux crate. For simplicity of use and implementation. The mapping is as follows:

lock behaves like a spinlock_irqsave. It returns a guard.

Dropping the guard behaves like spinlock_irqrestore

This means that locking a spinlock disables interrupts until all spinlock guards have been dropped.

A note on reordering: reordering lock drops is prohibited and doing so will result in UB.

Fields
internal: SpinLock<T>

SpinLock we wrap.

Implementations

impl<T> SpinLockIRQ<T>

pub const fn new(internal: T) -> SpinLockIRQ<T>

Creates a new spinlockirq wrapping the supplied data.

pub fn into_inner(self) -> T

Consumes this SpinLockIRQ, returning the underlying data.

impl<T: ?Sized> SpinLockIRQ<T>

pub fn lock(&self) -> SpinLockIRQGuard<T>

Disables interrupts and locks the mutex.

pub fn try_lock(&self) -> Option<SpinLockIRQGuard<T>>

Disables interrupts and locks the mutex.

pub unsafe fn force_unlock(&self)

Force unlocks the lock.

Safety

This is completely unsafe. It does not reset the interrupt status register, potentially causing deadlock. It should only be used when all hope is already lost.