Module sunrise_kernel::timer

The core timing of Sunrise.

Structs

IRQTimer
A stream of event that trigger every ns amount of nanoseconds, by counting interruptions.

KernelTimerInfo
This represent the information to derive all internal timing in Sunrise.

Statics

KERNEL_TIMER_INFO
Stores the information needed for Sunrise's internal timing.

Functions

set_kernel_timer_info
Set the information required for Sunrise timer to work.

wait_ns
Returns a stream of event that trigger every ns amount of nanoseconds.