Module sunrise_kernel::utils::loop_future

Definition of the LoopFn combinator, implementing Future loops.

Shamelessly stolen from the old futures.

Structs

LoopFn
A future implementing a tail-recursive loop.

Enums

Loop
The status of a loop_fn loop.

Poll
Indicates whether a value is available or if the current task has been scheduled to receive a wakeup instead.

Functions

loop_fn
Creates a new future implementing a tail-recursive loop.